requirements.txt
pyserial>=3.5
pywin32>=300; sys_platform == 'win32'



